+++ gradle-plugin/src/functionalTest/kotlin/dev/zacsweers/metro/gradle/incremental/testUtil.kt
@@ -22,6 +22,9 @@ private val DEFAULT_IMPORTS =
     // For Callable access
     "java.util.concurrent.*",
   )
+private val FILE_PATH_REGEX = Regex("file://.*?/(?=[^/]+\\.kt)")
+
+fun String.cleanOutputLine(): String = FILE_PATH_REGEX.replace(trimEnd(), "")
 
 fun GradleProject.classLoader(): ClassLoader {
   val classesDir = rootDir.toPath().resolve("build/classes/kotlin/main").absolute()